fix: BF16 dtype compatibility across RMSNorm, LinearAttnNode, TiedLMHead, Trainer AMP
- RMSNorm: cast output back to input dtype (was leaking FP32 downstream)
- LinearAttnNode: cast q_feat/k_feat weights to FP32 for the matmul, keeping output FP32 naturally
- TiedLMHead: defensive dtype alignment + cast weight to h_mixed.dtype
- Trainer: add amp_dtype parameter, default from cfg.amp_dtype, use in torch.amp.autocast instead of hardcoded float16
- Config: add amp_dtype field (float16 | bfloat16)

+++ helix_lm/hf_model.py
@@ -74,6 +74,8 @@ class TiedLMHead(nn.Module):
         # h: (B, T, d_model)
         # P0 FIX: Buffer applied consistently in BOTH train and eval.
         # Forward pass must NEVER depend on self.training.
+        # Defensive: align h dtype to buffer weight dtype (guards against upstream casts)
+        h = h.to(self.buffer.weight.dtype)
         if 0 < self.grad_buffer_ratio < 1:
             h_buffered = self.buffer(h)
             h_mixed = (1 - self.grad_buffer_ratio) * h + self.grad_buffer_ratio * h_buffered
@@ -82,7 +84,7 @@ class TiedLMHead(nn.Module):
         else:
             # buffer_ratio=0: pass through directly (standard tying)
             h_mixed = h
-        return F.linear(h_mixed, self.weight)
+        return F.linear(h_mixed, self.weight.to(h_mixed.dtype))
 
 
 # ---------------------------------------------------------------------------
